requirements.txt
cffi==1.9.1
cryptography==1.7.1
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.10
django-bootstrap3==7.0.1
django-braces==1.10.0
django-oauth-toolkit==0.11.0
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.4.3
gunicorn==19.6.0
idna==2.2
ndg-httpsclient==0.4.2
oauthlib==1.1.2
Pillow==3.3.0
pyasn1==0.1.9
pycparser==2.17
PyJWT==1.4.2
pyOpenSSL==16.2.0
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2016.6.1
requests==2.12.3
requests-oauthlib==0.7.0
six==1.10.0
stripe==1.37.0
urllib3==1.19.1
whitenoise==3.2.1
